### 2025-12-08 #9574 @editorialbot
[PRE REVIEW]: OMEinsumContractionOrders: A Julia package for tensor network contraction order optimization
**Submitting author:** @GiggleLiu (<a href="http://orcid.org/0000-0003-1635-2679">Jin-Guo Liu</a>)
**Repository:** https://github.com/TensorBFS/OMEinsumContractionOrders.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.2.3
