+++ paper/paper.bib
@@ -151,7 +151,8 @@
   number={90},
   pages={297--301},
   year={1965},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi={10.7551/mitpress/5222.003.0014}
 }
 
 @article{Alman2024,
@@ -377,6 +378,7 @@
   year = {2019},
   eprint = {1902.03178v3},
   pages = {1--28},
+  doi = {10.22331/q-2020-06-04-279},
   archiveprefix = {arXiv},
   file = {/home/leo/snap/zotero-snap/common/Zotero/storage/VPGJQY2D/Duncan et al_2019_Graph-theoretic Simplification of Quantum Circuits with the ZX-calculus.pdf}
 }
@@ -530,7 +532,8 @@
   number={2},
   pages={1098},
   year={1996},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/physreva.54.1098}
 }
 @book{gaitan2008quantum,
   title={Quantum error correction and fault tolerant quantum computing},
@@ -556,7 +559,8 @@
   number={1954},
   pages={2551--2577},
   year={1996},
-  publisher={The Royal Society London}
+  publisher={The Royal Society London},
+  doi={10.1098/rspa.1996.0136}
 }
 @article{kitaev2003fault,
   title={Fault-tolerant quantum computation by anyons},
@@ -566,7 +570,8 @@
   number={1},
   pages={2--30},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/s0003-4916(02)00018-0}
 }
 @article{dennis2002topological,
   title={Topological quantum memory},
@@ -594,7 +599,7 @@
   pages={A1239--A1270},
   year={2023},
   publisher={SIAM},
-  url={https://doi.org/10.1137/22M1501787}
+  doi={10.1137/22m1501787}
 }
 
 @article{Bouchitte2001,
@@ -606,7 +611,7 @@
   pages={212--232},
   year={2001},
   publisher={SIAM},
-  url={https://doi.org/10.1137/S0097539799359683}
+  doi={10.1137/s0097539799359683}
 }
 
 @article{Luo2020,
@@ -616,8 +621,8 @@
   volume={4},
   pages={341},
   year={2020},
-  url={https://doi.org/10.22331/q-2020-10-11-341},
-  publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften}
+  publisher={Verein zur F{\"o}rderung des Open Access Publizierens in den Quantenwissenschaften},
+  doi={10.22331/q-2020-10-11-341}
 }
 
 @article{Roa2023,
@@ -629,7 +634,7 @@
   pages={5700},
   year={2023},
   publisher={Open Journals},
-  url={https://joss.theoj.org/papers/10.21105/joss.05700}
+  doi={10.21105/joss.05700}
 }
 
 @article{Magron2021,
@@ -665,7 +670,7 @@
   pages={141--153},
   year={1985},
   publisher={ACM New York, NY, USA},
-  url={https://doi.org/10.1145/214392.214398}
+  doi={10.1145/214392.214398}
 }
 
 @article{Rothberg1998,
@@ -677,7 +682,7 @@
   pages={682--695},
   year={1998},
   publisher={SIAM},
-  url={https://doi.org/10.1137/S0895479896302692}
+  doi={10.1137/s0895479896302692}
 }
 
 @article{Arute2019,
