+++ paper/paper.bib
@@ -443,7 +443,7 @@
   eprint = {1805.00475},
   abstract = {We introduce tensor network contraction algorithms for the counting of satisfying assignments in constraint satisfaction problems ({\textbackslash}\#CSP). We represent each arbitrary {\textbackslash}\#CSP instance as a tensor network whose full contraction yields the number of satisfying assignments for that instance. We then use methods of graph theory and computational complexity theory to develop fast tensor contraction protocols. In particular, we employ the tools of multilevel graph partitioning and community structure detection to determine favorable orders of contraction of arbitrary tensor networks. We first prove analytically that full tensor contraction can be performed in subexponential time for {\textbackslash}\#CSP instances defined on graphs with sublinear separators. We then implement numerical heuristics for the solution of general {\textbackslash}\#P-hard counting boolean satisfiability ({\textbackslash}\#SAT) problems, focusing on random instances of {\textbackslash}\#{\textbackslash}textsc\{Cubic-Vertex-Cover\} as a concrete example, and show that they outperform state-of-the-art {\textbackslash}\#SAT solvers by a significant margin. Our results promote tensor network contraction algorithms as a powerful practical tool for fast solution of some {\textbackslash}\#CSPs.},
   archiveprefix = {arXiv},
-  file = {/Users/liujinguo/Zotero/storage/F5QSIPMX/Kourtis et al_2018_Fast counting with tensor networks.pdf}
+  doi = {10.21468/scipostphys.7.5.060}
 }
 
 @article{Schollwock2011,
