+++ mcp_server/pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "tensor-network-optimizer-mcp"
+version = "0.1.0"
+description = "MCP Server for Tensor Network Contraction Order Optimization"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+    "mcp>=1.0.0",
+]
+
+[dependency-groups]
+dev = []

[…]
