### 2026-01-08 #9574 @jolars
As you may have seen @GiggleLiu, JOSS has recently updated its scope and requirements. Please that your paper will be reviewed under the new requirements, which for instance expects additional sections in your paper. Please read [the announcement](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future) and the [updated requirements](https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance).
