+++ mcp_server/server.jl
@@ -1,0 +1,419 @@
+#!/usr/bin/env julia
+"""
+Pure Julia MCP Server for Tensor Network Contraction Order Optimization.
+
+This server provides tools for optimizing tensor network contraction orders
+using OMEinsumContractionOrders.jl directly, with no subprocess overhead.
+
+Usage:
+    julia --project=. server.jl
+
+For Cursor MCP configuration (~/.cursor/mcp.json):
+{
+  "mcpServers": {
+    "tensor-network-optimizer": {
+      "command": "julia",
+      "args": ["--project=/path/to/mcp_server", "/path/to/mcp_server/server.jl"]
+    }
+  }
+}
+"""
+
+using ModelContextProtocol
+using OMEinsumContractionOrders
+using JSON3
+
+# ============================================================================
+# Helper Functions
+# ============================================================================
+
+"""
+Parse label type from inputs - determines if labels are Char or Int.
+"""
+function parse_label_type(inputs::Vector, output::Vector)
+    all_labels = vcat(inputs..., output)
+    isempty(all_labels) && return Int
+    first_label = first(all_labels)
+    if first_label isa String && length(first_label) == 1
+        return Char
+    else
+        return Int
+    end
+end
+
+"""
+Convert labels to the target type.
+"""
+function convert_labels(inputs::Vector, output::Vector, ::Type{LT}) where LT
+    conv = if LT == Char
+        l -> l isa String ? only(l) : Char(l)
+    elseif LT == Int
+        l -> l isa String ? parse(Int, l) : Int(l)
+    else
+        l -> string(l)
+    end
+    converted_inputs = [[conv(l) for l in ix] for ix in inputs]
+    converted_output = [conv(l) for l in output]
+    return converted_inputs, converted_output
+end
+
+"""
+Parse size dictionary with proper label types.
+"""
+function parse_size_dict(size_dict::Dict, ::Type{LT}) where LT
+    conv = if LT == Char
+        k -> k isa String && length(k) == 1 ? only(k) : Char(k)
+    elseif LT == Int
+        k -> k isa String ? parse(Int, k) : Int(k)
+    else
+        k -> string(k)
+    end
+    return Dict{LT,Int}(conv(k) => Int(v) for (k, v) in size_dict)
+end
+
+"""
+Create optimizer from type string and parameters.
+"""
+function create_optimizer(optimizer_type::String, params::Dict)
+    opt = lowercase(optimizer_type)
+    
+    if opt == "greedy"
+        α = get(params, "alpha", get(params, "α", 0.0))
+        temperature = get(params, "temperature", 0.0)
+        return GreedyMethod(; α=α, temperature=temperature)
+        
+    elseif opt == "treesa"
+        sc_target = get(params, "sc_target", 20.0)
+        ntrials = get(params, "ntrials", 10)
+        niters = get(params, "niters", 50)
+        score = ScoreFunction(sc_target=sc_target)
+        return TreeSA(; ntrials=ntrials, niters=niters, score=score)
+        
+    elseif opt == "exacttreewidth" || opt == "exact_treewidth"
+        return ExactTreewidth()
+        
+    elseif opt == "treewidth"
+        return Treewidth()
+        
+    elseif opt == "sabipartite"
+        sc_target = get(params, "sc_target", 25)
+        ntrials = get(params, "ntrials", 50)
+        niters = get(params, "niters", 1000)
+        return SABipartite(; sc_target=sc_target, ntrials=ntrials, niters=niters)
+        
+    else
+        error("Unknown optimizer: $optimizer_type. Options: greedy, treesa, treewidth, exacttreewidth, sabipartite")
+    end
+end
+
+"""
+Convert NestedEinsum/SlicedEinsum to dictionary using built-in JSON support.
+"""
[…]
