+++ mcp_server/Project.toml
@@ -1,0 +1,7 @@
+[deps]
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
+ModelContextProtocol = "c58f755f-f2a7-4f48-bf29-4e9659b78499"
+OMEinsumContractionOrders = "6f22d1fd-8eed-4bb7-9776-e7d684900715"
+
+[sources]
+OMEinsumContractionOrders = {path = ".."}

[…]
