### 2026-01-09 c7c31be
Add required JOSS sections for updated submission requirements
Add Software Design section highlighting first open-source TreeSA implementation and build vs. contribute justification for native Julia implementation. Add Research Impact Statement with performance benchmarks and ecosystem adoption evidence including publications. Add AI Usage Disclosure section. Update bibliography with recent publications using OMECO.

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude Sonnet 4.5 <noreply@anthropic.com>
