+++ paper/paper.bib
@@ -696,3 +696,27 @@
   publisher={Nature Publishing Group UK London},
   url={https://www.nature.com/articles/s41586-019-1666-5}
 }
+
+@article{Gao2024crossentropy,
+  title={Limitations of Linear Cross-Entropy as a Measure for Quantum Advantage},
+  author={Gao, Xun and Kalinowski, Marcin and Chou, Chi-Ning and Lukin, Mikhail D and Barak, Boaz and Choi, Soonwon},
+  journal={PRX Quantum},
+  volume={5},
+  number={1},
+  pages={010334},
+  year={2024},
+  publisher={APS},
+  doi={10.1103/PRXQuantum.5.010334}
+}
+
+@article{Ebadi2022,
+  title={Quantum optimization of maximum independent set using Rydberg atom arrays},
+  author={Ebadi, Sepehr and Keesling, Alexander and Cain, Madelyn and Wang, Tout T and Levine, Harry and Bluvstein, Dolev and Semeghini, Giulia and Omran, Ahmed and Liu, Jin-Guo and Samajdar, Rhine and others},
+  journal={Science},
+  volume={376},
+  number={6598},
+  pages={1209--1215},
+  year={2022},
+  publisher={American Association for the Advancement of Science},
+  doi={10.1126/science.abo6587}
+}
