### 2026-01-09 #9574 @danielskatz
(an editorialbot command needs to be the first thing in a new comment)
