### 2026-01-18 #9574 @GiggleLiu
> Thanks, [@danielskatz](https://github.com/danielskatz)!
>
> [@GiggleLiu](https://github.com/GiggleLiu), do you possibly have any additional suggestions for reviewers for your submission? You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.

Sorry for the late reply, how about:
- shinaoka (author of tensor4all-rs)
