### 2026-01-23 #9886 @jolars
:wave: @meandmytram and @epatters - Thanks for agreeing to review this submission. This is the review thread for the paper. All of our communications will happen here from now on.

As you can see above, you each should use the command

```@editorialbot generate my checklist```

to create your review checklist. @editorialbot commands need to be the first thing in a new comment.
