### 2026-01-23 #9886 @jolars
@GiggleLiu, your paper seems to be missing the required **State of the field** section:

> State of the field: A description of how this software compares to other commonly-used packages in the research area. If related tools exist, provide a clear “build vs. contribute” justification explaining your unique scholarly contribution and why existing alternatives are insufficient.
