### 2026-02-03 #9886 @jolars
:wave: @GiggleLiu, your paper seems to be missing the required State of the field section (as noted above). Can you please add it and regenerate the PDF afterwards?
