### 2026-02-04 de0a49a
add state of the field
