### 2026-02-05 5f182da
Address review comments on State of the field section
- Reformat Smith2018 and Gray2018 bib entries to multi-line style
- Fix author name capitalization: "Daniel G. a." → "Daniel G. A."
- Remove redundant "widely used" phrasing for cotengra

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
