+++ paper/paper.bib
@@ -733,9 +733,31 @@
 	url={https://scipost.org/10.21468/SciPostPhysCodeb.4},
 }
 
-@article{Smith2018, doi = {10.21105/joss.00753}, url = {https://doi.org/10.21105/joss.00753}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {26}, pages = {753}, author = {Smith, Daniel G. a. and Gray, Johnnie}, title = {opt\_einsum - A Python package for optimizing contraction order for einsum-like expressions}, journal = {Journal of Open Source Software} }
+@article{Smith2018,
+	title = {opt\_einsum - A Python package for optimizing contraction order for einsum-like expressions},
+	author = {Smith, Daniel G. A. and Gray, Johnnie},
+	journal = {Journal of Open Source Software},
+	volume = {3},
+	number = {26},
+	pages = {753},
+	year = {2018},
+	publisher = {The Open Journal},
+	doi = {10.21105/joss.00753},
+	url = {https://doi.org/10.21105/joss.00753}
+}
 
-@article{Gray2018, doi = {10.21105/joss.00819}, url = {https://doi.org/10.21105/joss.00819}, year = {2018}, publisher = {The Open Journal}, volume = {3}, number = {29}, pages = {819}, author = {Gray, Johnnie}, title = {quimb: A python package for quantum information and many-body calculations}, journal = {Journal of Open Source Software} }
+@article{Gray2018,
+	title = {quimb: A python package for quantum information and many-body calculations},
+	author = {Gray, Johnnie},
+	journal = {Journal of Open Source Software},
+	volume = {3},
+	number = {29},
+	pages = {819},
+	year = {2018},
+	publisher = {The Open Journal},
+	doi = {10.21105/joss.00819},
+	url = {https://doi.org/10.21105/joss.00819}
+}
 
 @software{TensorOperations,
 	author  = {Lukas Devos and Maarten Van Damme and Jutho Haegeman and contributors},
