+++ paper/paper.md
@@ -86,7 +86,7 @@ These applications are reflected in the ecosystem built around OMECO, as illustr
 
 # State of the field
 
-Several software packages support contraction order optimization. The Python package `cotengra` [@Gray2021] is widely used and offers multiple strategies, including greedy heuristics and simulated annealing. However, because `cotengra` is implemented in Python, integrating it into Julia-first scientific computing workflows can be less convenient and may introduce additional overhead in performance-critical pipelines. Other packages, such as `opt_einsum` [@Smith2018] and `quimb` [@Gray2018], offer contraction order optimization within broader numerical and tensor network ecosystems, but typically emphasize a narrower set of heuristics or provide less comprehensive algorithmic coverage. In the Julia ecosystem, `TensorOperations.jl` [@TensorOperations] and `ITensors.jl` [@ITensor] primarily focus on efficient tensor operations and physics applications, rather than dedicated contraction order optimization.
+Several software packages support contraction order optimization. The Python package `cotengra` [@Gray2021] provides multiple optimization strategies, including greedy heuristics and simulated annealing. However, because `cotengra` is implemented in Python, integrating it into Julia-first scientific computing workflows can be less convenient and may introduce additional overhead in performance-critical pipelines. Other packages, such as `opt_einsum` [@Smith2018] and `quimb` [@Gray2018], offer contraction order optimization within broader numerical and tensor network ecosystems, but typically emphasize a narrower set of heuristics or provide less comprehensive algorithmic coverage. In the Julia ecosystem, `TensorOperations.jl` [@TensorOperations] and `ITensors.jl` [@ITensor] primarily focus on efficient tensor operations and physics applications, rather than dedicated contraction order optimization.
 
 # Software Design
 
