### 2026-02-05 8df3db1
Improve State of the field section accuracy
- Reformat Smith2018 and Gray2018 bib entries to multi-line style
- Fix author name: "Daniel G. a." → "Daniel G. A."
- Clarify cotengra as hyper-optimization framework (not just greedy/SA)
- Fix quimb description: uses cotengra as backend
- Fix opt_einsum: includes dynamic programming, not just greedy
- Fix TensorOperations.jl: exact optimization, no hyperedges, limited scalability
- Fix ITensors.jl: focus on MPS/DMRG physics applications
- Fix Python integration issue: version control, not performance

Co-Authored-By: Claude Opus 4.5 <noreply@anthro
[…]
