### 2026-02-05 8d930d5
add state of the field (#105)
* add state of the field

* Address review comments on State of the field section

- Reformat Smith2018 and Gray2018 bib entries to multi-line style
- Fix author name capitalization: "Daniel G. a." → "Daniel G. A."
- Remove redundant "widely used" phrasing for cotengra

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>

* Improve State of the field section accuracy

- Reformat Smith2018 and Gray2018 bib entries to multi-line style
- Fix author name: "Daniel G. a." → "Daniel G. A."
- Clarify cotengra as hyper-optimization framework (not just greedy/SA)
- Fix 
[…]
