+++ paper/paper.bib
@@ -720,3 +720,51 @@
   publisher={American Association for the Advancement of Science},
   doi={10.1126/science.abo6587}
 }
+
+
+@article{ITensor,
+	title={{The ITensor Software Library for Tensor Network Calculations}},
+	author={Matthew Fishman and Steven R. White and E. Miles Stoudenmire},
+	journal={SciPost Phys. Codebases},
+	pages={4},
+	year={2022},
+	publisher={SciPost},
+	doi={10.21468/SciPostPhysCodeb.4},
+	url={https://scipost.org/10.21468/SciPostPhysCodeb.4},
+}
+
+@article{Smith2018,
+	title = {opt\_einsum - A Python package for optimizing contraction order for einsum-like expressions},
+	author = {Smith, Daniel G. A. and Gray, Johnnie},
+	journal = {Journal of Open Source Software},
+	volume = {3},
+	number = {26},
+	pages = {753},
+	year = {2018},
+	publisher = {The Open Journal},
+	doi = {10.21105/joss.00753},
+	url = {https://doi.org/10.21105/joss.00753}
+}
+
+@article{Gray2018,
+	title = {quimb: A python package for quantum information and many-body calculations},
+	author = {Gray, Johnnie},
+	journal = {Journal of Open Source Software},
+	volume = {3},
+	number = {29},
+	pages = {819},
+	year = {2018},
+	publisher = {The Open Journal},
+	doi = {10.21105/joss.00819},
+	url = {https://doi.org/10.21105/joss.00819}
+}
+
+@software{TensorOperations,
+	author  = {Lukas Devos and Maarten Van Damme and Jutho Haegeman and contributors},
+	title   = {TensorOperations.jl},
+	url     = {https://github.com/Jutho/TensorOperations.jl},
+	version = {v4.0.7},
+	doi		= {10.5281/zenodo.3245496},
+	year    = {2023},
+	month   = {10}
+}
