### 2026-02-10 #9886 @meandmytram
hey @jolars I am in process of reproducing the plots presented in the paper as per JOSS guidelines it seems that they are considered original results
