### 2026-02-11 #9886 @meandmytram
hi everyone, I was able to reproduce the aforementioned results, I don't have any other points left thus I can recommend the package for publicaiton, cheers
