### 2026-03-01 #9886 @epatters
## Review comments

Apologies for the delay in returning my review. The software appears to be high quality and the paper well written. I would recommend it for publication in JOSS. A few specific comments below.

> * **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

The standard GitHub infrastructure goes some way toward this, but adding a `CONTRIBUTING.md` is a standard way to convey guidelines for new contributors.

> * **AI usage disclosure (required section):** Does the paper include an "AI usage disclosure" section that transparently discloses any use of generative AI in the software creation, documentation, or paper authoring?

The authors acknowledge the use of LLM assistance for writing, and the paper does indeed show signs of LLM use, including at times the notorious feature of excessive and unnecessary usage of hyphens. While tastes may vary, I personally find it distracting to think about such things when reading a document. For the future, I would suggest to the authors that they will be better served by finding their own authorial voice, imperfect and idiosyncratic though it may be.

> * **State of the field (required section):** Do the authors describe how this software compares to other commonly-used packages? If related tools exist, do they provide a clear "build vs. contribute" justification explaining their unique scholarly contribution and why alternatives are insufficient?

AFAIU, a key differentiating feature of this package compared to existing packages is stated in this sentence: "OMECO provides the first open-source implementation of the TreeSA algorithm, a powerful simulated annealing approach that operates directly on contraction tree structures." However, that sentence and accompanying paragraph appear in the section "Software Design." It seems to me that the paragraph should be in either "Statement of need" or "State of the field."

> * **Software Design (required section):** Does the paper include a "Software Design" section that explains the trade-offs weighed, the design/architecture chosen, and why it matters? Is the content compelling and does it demonstrate meaningful design thinking?

Assuming one moves the first paragraph of the "Software design" section elsewhere, as suggested above, there is then only one paragraph in the manuscript about software design. I expect that this section could be usefully expanded.

For example, the authors write: "OMECO’s architecture emphasizes modularity and extensibility. The core abstraction separates the contraction order representation (as trees or paths) from the optimization algorithms, enabling users to compose different strategies and implement custom optimizers". How does it achieve that? What does it mean to compose different strategies? In general, these are not straightforward matters.

> * **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The paper is generally clear and readable.

The formatting of the section titles is currently inconsistent: some section titles have each word capitalized, and others have only the first word capitalized.
