+++ paper/paper.md
@@ -59,7 +59,7 @@ OMECO is designed to search for near-optimal contraction orders for exact tensor
 In this paper, we present the key features of OMECO, its integration with the Julia ecosystem, and performance benchmarks.
 
 
-# Statement of need
+# Statement of Need
 
 A _tensor network_ is a mathematical structure that represents multilinear transformations as hypergraphs. Arrays—called _tensors_—correspond to nodes, and shared indices correspond to hyperedges. To _contract_ a tensor network is to evaluate the transformation on a collection of tensors by performing a sequence of pairwise bilinear operations. The computational cost—both running time and memory usage—depends critically on the order in which these operations are performed. A specific choice of ordering is called a _contraction order_, and the problem of finding an efficient ordering is called _contraction order optimization_.
 
@@ -84,15 +84,19 @@ These applications are reflected in the ecosystem built around OMECO, as illustr
 
 ![The ecosystem built around `OMEinsumContractionOrders` and its dependencies. OMECO serves as a core component of the tensor network contractor `OMEinsum`, which powers applications including `Yao` (quantum simulation), `TensorQEC` (quantum error correction), `TensorInference` (probabilistic inference), `GenericTensorNetworks` and `TensorBranching` (combinatorial optimization).\label{fig:structure}](figures/structure.pdf){ width=80% }
 
-# State of the field
+# State of the Field
 
 Several software packages support contraction order optimization. The Python package `cotengra` [@Gray2021] provides a hyper-optimization framework that combines greedy methods, simulated annealing, and graph partitioning to find high-quality contraction orders. However, integrating Python packages into Julia workflows introduces challenges in dependency version control and environment reproducibility. The `opt_einsum` package [@Smith2018] provides contraction optimization for einsum-style expressions using dynamic programming and greedy-based heuristics, while `quimb` [@Gray2018] offers tensor network capabilities using `cotengra` as its optimization backend. In the Julia ecosystem, `TensorOperations.jl` [@TensorOperations] implements exact contraction order optimization, but is limited to tensor networks without hyperedges (where each index appears in at most two tensors) and lacks scalability to large networks. `ITensors.jl` [@ITensor] primarily focuses on physics applications such as MPS and DMRG algorithms, rather than general-purpose contraction order optimization.
 
+Different from most of these exsiting packages, OMECO focus on optimizing arbitrary tensor networks with hyperedges, and provides the first open-source implementation of the TreeSA algorithm [@Kalachev2021], a powerful simulated annealing approach that operates directly on contraction tree structures. Beyond TreeSA, OMECO implements a comprehensive suite of optimization algorithms from the literature, including nested dissection methods for hypergraphs, bipartition-based approaches, and exact tree decomposition solvers. While the Python package `cotengra` [@Gray2021] exists in this space, we chose to build OMECO as a native Julia implementation to provide these capabilities to the Julia scientific computing ecosystem without language barriers, enabling new applications in quantum computing and combinatorial optimization that require tight integration with Julia's tensor network infrastructure.
+
 # Software Design
 
-OMECO provides the first open-source implementation of the TreeSA algorithm [@Kalachev2021], a powerful simulated annealing approach that operates directly on contraction tree structures. Beyond TreeSA, OMECO implements a comprehensive suite of optimization algorithms from the literature, including nested dissection methods for hypergraphs, bipartition-based approaches, and exact tree decomposition solvers. While the Python package `cotengra` [@Gray2021] exists in this space, we chose to build OMECO as a native Julia implementation to provide these capabilities to the Julia scientific computing ecosystem without language barriers, enabling new applications in quantum computing and combinatorial optimization that require tight integration with Julia's tensor network infrastructure.
[…]
