### 2026-03-08 TensorBFS/OMEinsumContractionOrders.jl#108 @xuanzhaogao
Do we need a contributing.md?
As suggested by https://github.com/openjournals/joss-reviews/issues/9886#issuecomment-3980794355
