### 2026-03-12 0d5fa83
Update paper/paper.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
