### 2026-03-12 b5e2e0a
revise paper.md (#107)
* add state of the field

* Address review comments on State of the field section

- Reformat Smith2018 and Gray2018 bib entries to multi-line style
- Fix author name capitalization: "Daniel G. a." → "Daniel G. A."
- Remove redundant "widely used" phrasing for cotengra

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>

* Improve State of the field section accuracy

- Reformat Smith2018 and Gray2018 bib entries to multi-line style
- Fix author name: "Daniel G. a." → "Daniel G. A."
- Clarify cotengra as hyper-optimization framework (not just greedy/SA)
- Fix quimb d
[…]
