+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"background/#Tensor-Network-Contraction-Order-Optimization","page":"Background Knowledge","title":"Tensor Network Contraction Order Optimization","text":"","category":"section"},{"location":"background/#Tensor-network","page":"Background Knowledge","title":"Tensor network","text":"","category":"section"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"Tensor network is a powerful tool for modeling and simulating quantum many-body systems, probabilistic inference, combinatorial optimization, etc. It is a diagrammatic representation of tensor contractions. In this representation, a tensor is represented as a node, and an index is represented as a hyperedge (a hyperedge can connect to any number of nodes). For example, vectors, matrices and higher order tensors can be represented as:","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"(Image: )","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"To understand the basic concepts of tensor network, we recommend the following references for readers with different background:","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"For readers with physics background: https://tensornetwork.org/diagrams/\nFor readers want to get a formal definition: Chapter 2 of https://epubs.siam.org/doi/abs/10.1137/22M1501787","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"The defining operation of a tensor network is contraction, which is also named as sum-product operation. It can be viewed as the generalization of matrix multiplication to tensors. We illustrate tensor network contraction with the following example.","category":"page"},{"location":"background/#Example:-Trace-permutation-rule","page":"Background Knowledge","title":"Example: Trace permutation rule","text":"","category":"section"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"Let A B and C be three square matrices with the same size. The trace operation texttr(A B C) is defined as the sum of the products of the elements of the matrices:","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"texttr(A B C) = sum_ijk A_ij B_ik C_jk","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"It can be represented as a tensor network diagram as follows:","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"(Image: )","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"So, each tensor is represented as a node, and each index is represented as an edge. Tensors sharing the same index are connected by the edge. The contraction of the tensor network is the summation of the products of the elements of the tensors over all possible values of the indices.","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"The contraction can happen in different orders. The famous trace permutation rule states that the trace is invariant under cyclic permutations of the matrices:","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"texttr(A B C) = texttr(C A B) = texttr(B C A)","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"Hence, we observe that tensor network contraction is associative and commutative, i.e., the order of evaluation does not change the result. In the tensor network diagram, the order of evaluation is neglected, so the trace permutation rule is trivially satisfied.","category":"page"},{"location":"background/","page":"Background Knowledge","title":"Background Knowledge","text":"By drawing diagrams, we prove the trace permutation rule. Isn't it coo
[…]
