### 2026-03-12 #9886 @GiggleLiu
Thank you @jolars and @epatters for your time. To address the issues, changes we have made:

1. Add "Community" guidelines. I agree it is nessesary. Instead of creating a `CONTRIBUTING.md`, I created a `CLAUDE.md` and `AGENTS.md` instead (mentioned in README Contributor Guide section). It serve for the same purpose of guiding contributors to contribute, but can also be read by AI automatically.

2. Improved writing. We adjusted the writing of "State of the field" and "Software Design" accordingly, and made the title consistent.
