### 2026-03-16 #9886 @jolars
> [@editorialbot](https://github.com/editorialbot) generate pdf
>
> Thank you [@jolars](https://github.com/jolars) and [@epatters](https://github.com/epatters) for your time. To address the issues, changes we have made:
>
>     1. Add "Community" guidelines. I agree it is nessesary. Instead of creating a `CONTRIBUTING.md`, I created a `CLAUDE.md` and `AGENTS.md` instead (mentioned in README Contributor Guide section). It serve for the same purpose of guiding contributors to contribute, but can also be read by AI automatically.
>
>     2. Improved writing. We adjusted the writing of "State of the field" and "Software Design" accordingly, and made the title consistent.

:wave: @epatters - are you happy with these changes?
