+++ paper/paper.bib
@@ -615,7 +615,7 @@
 }
 
 @article{Luo2020,
-  title={Yao. jl: Extensible, efficient framework for quantum algorithm design},
+  title={{{Yao.jl}}: Extensible, efficient framework for quantum algorithm design},
   author={Luo, Xiu-Zhe and Liu, Jin-Guo and Zhang, Pan and Wang, Lei},
   journal={Quantum},
   volume={4},
