### 2026-03-23 c79fa19
Increase size of figure
