### 2026-03-23 b7bbcfd
Add hyphen between "width" and "based"
