### 2026-03-24 e374809
JOSS paper comments (#114)
* Protect `Yao.jl` and remove etraneous space

* Add Oxford comma for consistency

* Flip package order to mitigate escape into margin

* Increase size of figure

* Change "The" to "A"

* Add hyphen between "width" and "based"

* Add missing article before y-axis
