+++ paper/paper.bib
@@ -23,34 +23,14 @@
   doi = {10.48550/arXiv.1209.5145},
   abstract = {Dynamic languages have become popular for scientific computing. They are generally considered highly productive, but lacking in performance. This paper presents Julia, a new dynamic language for technical computing, designed for performance from the beginning by adapting and extending modern programming language techniques. A design based on generic functions and a rich type system simultaneously enables an expressive programming model and successful type inference, leading to good performance for a wide range of programs. This makes it possible for much of the Julia library to be written in Julia itself, while also incorporating best-of-breed C and Fortran libraries.},
   urldate = {2018-11-27},
-  journal = {arXiv:1209.5145 [cs]},
+  journal = {arXiv preprint arXiv:1209.5145},
   author = {Bezanson, Jeff and Karpinski, Stefan and Shah, Viral B. and Edelman, Alan},
   month = sep,
   year = {2012},
-  note = {arXiv: 1209.5145},
+  eprint = {1209.5145},
+  archiveprefix = {arXiv},
   keywords = {Computer Science - Programming Languages, Computer Science - Computational Engineering, Finance, and Science, D.3.2},
   file = {arXiv.org Snapshot:/Users/apodusenko/Zotero/storage/V6P4IU54/1209.html:text/html;Bezanson et al. - 2012 - Julia A Fast Dynamic Language for Technical Compu.pdf:/Users/apodusenko/Zotero/storage/4Q9THC8V/Bezanson et al. - 2012 - Julia A Fast Dynamic Language for Technical Compu.pdf:application/pdf},
-}
-
-@article{Golub2016,
-  title = {Matrix {{Computation}}},
-  author = {Golub, Gene H.},
-  editor = {{Intergovernmental Panel on Climate Change}},
-  year = {2016},
-  volume = {25},
-  number = {3},
-  eprint = {1011.1669v3},
-  pages = {228--234},
-  publisher = {Cambridge University Press},
-  address = {Cambridge},
-  issn = {10623264},
-  doi = {10.4037/ajcc2016979},
-  abstract = {Predicting the binding mode of flexible polypeptides to proteins is an important task that falls outside the domain of applicability of most small molecule and protein-protein docking tools. Here, we test the small molecule flexible ligand docking program Glide on a set of 19 non-{$\alpha$}-helical peptides and systematically improve pose prediction accuracy by enhancing Glide sampling for flexible polypeptides. In addition, scoring of the poses was improved by post-processing with physics-based implicit solvent MM- GBSA calculations. Using the best RMSD among the top 10 scoring poses as a metric, the success rate (RMSD {$\leq$} 2.0 {\AA} for the interface backbone atoms) increased from 21\% with default Glide SP settings to 58\% with the enhanced peptide sampling and scoring protocol in the case of redocking to the native protein structure. This approaches the accuracy of the recently developed Rosetta FlexPepDock method (63\% success for these 19 peptides) while being over 100 times faster. Cross-docking was performed for a subset of cases where an unbound receptor structure was available, and in that case, 40\% of peptides were docked successfully. We analyze the results and find that the optimized polypeptide protocol is most accurate for extended peptides of limited size and number of formal charges, defining a domain of applicability for this approach.},
-  archiveprefix = {arXiv},
-  isbn = {9788578110796},
-  pmid = {25246403},
[…]
