### 2026-03-24 0b6a108
Verify and correct paper reference metadata
