### 2026-03-24 da147d2
Address JOSS issues 109-113 in paper references (#115)
* Address JOSS paper review issues 109-113

* Verify and correct paper reference metadata
