+++ paper/paper.md
@@ -111,7 +111,7 @@ The software exhibits strong community-readiness signals: comprehensive document
 # Features and Benchmarks
 
 A major feature of OMECO is contraction order optimization.
-OMECO provides several algorithms with complementary performance characteristics that can be simply called by the `optimize_code` function:
+OMECO provides several algorithms with complementary performance characteristics that can be simply called by the `optimize_code` function. \autoref{tbl:optimizers} summarizes the available optimizers and their roles:
 
 | Optimizer | Description |
 | :----------- | :------------- |
@@ -123,6 +123,8 @@ OMECO provides several algorithms with complementary performance characteristics
 | `SABipartite` | Simulated annealing bipartition method, pure Julia implementation |
 | `ExactTreewidth` | Exact algorithm with exponential runtime [@Bouchitte2001], based on [`TreeWidthSolver`](https://github.com/ArrogantGao/TreeWidthSolver.jl) |
 | `Treewidth` | Clique tree elimination methods from `CliqueTrees` package [@CliqueTrees2025] |
+
+: Summary of the contraction order optimizers available through `optimize_code`. {#tbl:optimizers}
 
 The algorithms `HyperND`, `Treewidth`, and `ExactTreewidth` are tree-width-based solvers that operate on graphs. They first convert tensor networks to their line graph representation [@Markov2008] and then find an optimized tree decomposition of the line graph using the `CliqueTrees` and `TreeWidthSolver` packages, as illustrated in \autoref{fig:structure}. Additionally, the `PathSA` optimizer optimizes path decomposition instead of tree decomposition. It is a variant of `TreeSA` by constraining contraction orders to path graphs, which is useful for applications requiring a linear contraction order.
 
@@ -178,4 +180,3 @@ The figure below illustrates these concepts with (a) a tensor network containing
 
 The tree decomposition in \autoref{fig:mainfig}(c) consists of 6 bags, each containing at most 3 indices, indicating that the treewidth of the tensor network is 2. The tensors $T_1$, $T_2$, $T_3$ and $T_4$ are contained in bags $B_1$, $B_5$, $B_6$ and $B_2$ respectively. Following the tree structure, we perform the contraction from the leaves. First, we contract bags $B_1$ and $B_2$ into $B_3$, yielding an intermediate tensor $I_{14} = T_1 * T_4$ (where "$*$" denotes tensor contraction) with indices $B$ and $E$. Next, we contract bags $B_5$ and $B_6$ into $B_4$, producing another intermediate tensor $I_{23} = T_2 * T_3$ also with indices $B$ and $E$. Finally, contracting $B_3$ and $B_4$ yields the desired scalar result.
  -->
-
