### 2026-03-24 d644ebd
Fix remaining bib entry issues from JOSS review (#116)
- Restore "Recursive" in Kalachev2021 title (was erroneously dropped)
- Update Pan2021 to published PRL version (Phys. Rev. Lett. 128, 030501, 2022)
- Update Magron2021 to published ACM TOMS version (vol. 49, no. 4, 2023)

Co-authored-by: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
