+++ paper/paper.bib
@@ -49,10 +49,15 @@
 }
 
 @article{Pan2021,
-      title={Simulating the Sycamore quantum supremacy circuits}, 
+      title={Simulating the Sycamore quantum supremacy circuits},
       author={Feng Pan and Pan Zhang},
-      journal={arXiv preprint arXiv:2103.03074},
-      year={2021},
+      journal={Physical Review Letters},
+      volume={128},
+      number={3},
+      pages={030501},
+      year={2022},
+      publisher={American Physical Society},
+      doi={10.1103/PhysRevLett.128.030501},
       eprint={2103.03074},
       archivePrefix={arXiv},
       primaryClass={quant-ph}
@@ -83,7 +88,7 @@
    pages={410}
 }
 @article{Kalachev2021,
-      title={Multi-Tensor Contraction for XEB Verification of Quantum Circuits},
+      title={Recursive Multi-Tensor Contraction for XEB Verification of Quantum Circuits},
       author={Gleb Kalachev and Pavel Panteleev and Man-Hong Yung},
       journal={arXiv preprint arXiv:2108.05665},
       year={2021},
@@ -621,13 +626,17 @@
 }
 
 @article{Magron2021,
-  title={TSSOS: a Julia library to exploit sparsity for large-scale polynomial optimization},
+  title={{TSSOS}: A Julia Library to Exploit Sparsity for Large-Scale Polynomial Optimization},
   author={Magron, Victor and Wang, Jie},
-  journal={arXiv preprint arXiv:2103.00915},
-  year={2021},
+  journal={ACM Transactions on Mathematical Software},
+  volume={49},
+  number={4},
+  pages={1--31},
+  year={2023},
+  publisher={ACM},
+  doi={10.1145/3618299},
   eprint={2103.00915},
-  archiveprefix={arXiv},
-  url={https://arxiv.org/abs/2103.00915}
+  archiveprefix={arXiv}
 }
 
 @misc{CliqueTrees2025,
