### 2026-03-24 #9886 @jolars
Thanks, @GiggleLiu ! I just see one more issue with something escaping into the margin. I tried to fix this in the PR but it didn't seem to work, maybe I have a different toolchain than editorial bot does. Can you try to see if you can fix this?
