### 2026-03-30 6464db0
Use correct table label syntax
