### 2026-03-30 #9886 @jolars
There were two issues with the paper, @GiggleLiu - I've filed https://github.com/TensorBFS/OMEinsumContractionOrders.jl/pull/117 to fix them.
