### 2026-03-30 33f88c6
Fix use of `\rm` and incorrect label syntax in paper (#117)
* Replace `\rm` with `\textrm{}`

* Use correct table label syntax
