+++ paper/paper.bib
@@ -49,7 +49,7 @@
 }
 
 @article{Pan2021,
-      title={Simulating the Sycamore quantum supremacy circuits},
+      title={Simulating the {S}ycamore quantum supremacy circuits},
       author={Feng Pan and Pan Zhang},
       journal={Physical Review Letters},
       volume={128},
@@ -248,7 +248,7 @@
 }
 
 @article{Markov2008,
-  title = {Simulating {{Quantum Computation}} by {{Contracting Tensor Networks}}},
+  title = {Simulating Quantum Computation by Contracting Tensor Networks},
   author = {Markov, Igor L. and Shi, Yaoyun},
   year = {2008},
   month = jan,
@@ -333,7 +333,7 @@
 }
 
 @inproceedings{Villescas2023,
-  title = {Scaling {{Probabilistic Inference Through Message Contraction Optimization}}},
+  title = {Scaling Probabilistic Inference Through Message Contraction Optimization},
   booktitle = {2023 {{Congress}} in {{Computer Science}}, {{Computer Engineering}}, \& {{Applied Computing}} ({{CSCE}})},
   author = {Villescas, Martin Roa and Liu, Jin-Guo and Wijnings, Patrick W.A. and Stuijk, Sander and Corporaal, Henk},
   year = {2023},
@@ -453,7 +453,7 @@
 }
 
 @article{Roa2024,
-  title = {Probabilistic {{Inference}} in the {{Era}} of {{Tensor Networks}} and {{Differential Programming}}},
+  title = {Probabilistic Inference in the Era of Tensor Networks and Differential Programming},
   author = {{Roa-Villescas}, Martin and Gao, Xuanzhao and Stuijk, Sander and Corporaal, Henk and Liu, Jin-Guo},
   year = {2024},
   month = sep,
@@ -486,7 +486,7 @@
 }
 
 @article{Piveteau2024,
-  title = {Tensor-{{Network Decoding Beyond 2D}}},
+  title = {Tensor-Network Decoding Beyond {2D}},
   author = {Piveteau, Christophe and Chubb, Christopher T. and Renes, Joseph M.},
   year = {2024},
   month = oct,
@@ -614,7 +614,7 @@
 }
 
 @article{Roa2023,
-  title={TensorInference: A Julia package for tensor-based probabilistic inference},
+  title={TensorInference: A {J}ulia package for tensor-based probabilistic inference},
   author={Roa-Villescas, Martin and Liu, Jin-Guo},
   journal={Journal of Open Source Software},
   volume={8},
@@ -626,7 +626,7 @@
 }
 
 @article{Magron2021,
-  title={{TSSOS}: A Julia Library to Exploit Sparsity for Large-Scale Polynomial Optimization},
+  title={{TSSOS}: A {J}ulia Library to Exploit Sparsity for Large-Scale Polynomial Optimization},
   author={Magron, Victor and Wang, Jie},
   journal={ACM Transactions on Mathematical Software},
   volume={49},
@@ -642,7 +642,7 @@
 @misc{CliqueTrees2025,
   author = {Samuelson, Richard and Fairbanks, James},
   url = {https://github.com/AlgebraicJulia/CliqueTrees.jl},
-  title = {CliqueTrees.jl: A Julia library for computing tree decompositions and chordal completions of graphs},
+  title = {CliqueTrees.jl: A {J}ulia library for computing tree decompositions and chordal completions of graphs},
   year = {2025}
 }
 
@@ -704,7 +704,7 @@
 }
 
 @article{Ebadi2022,
-  title={Quantum optimization of maximum independent set using Rydberg atom arrays},
+  title={Quantum optimization of maximum independent set using {R}ydberg atom arrays},
[…]
