### 2026-03-30 #9886 @danielskatz
@GiggleLiu - As track editor, I've proofread your paper and suggest the changes in https://github.com/TensorBFS/OMEinsumContractionOrders.jl/pull/118 - please merge this or let me know what you disagree with.

Also, please add countries to each affiliation.
