### 2026-03-31 2cdb6f3
Add countries to author affiliations per JOSS editor request
Co-Authored-By: Claude Opus 4.6 (1M context) <noreply@anthropic.com>
