### 2026-03-31 #9886 @danielskatz
Congratulations to @GiggleLiu and co-authors on your JOSS publication!!

JOSS completely depends on volunteers - we couldn't do this without you
